+++ tiago_lfc/launch/gz.py
@@ -5,6 +5,10 @@ from collections.abc import (
     Generator,
     Mapping,
 )
+from enum import (
+    Flag,
+    auto,
+)
 from pathlib import (
     Path,
 )
@@ -41,6 +45,15 @@ from .logging import (
 from .utils import (
     dict_to_string,
 )
+
+
+def __bool_from_string(s: Text) -> Optional[bool]:
+    if s == 'True':
+        return True
+    elif s == 'False':
+        return False
+    else:
+        return None
 
 
 def __log_then_forward_cmd(cmd):
@@ -128,7 +141,7 @@ def gz_server(
             choices=['False', 'True'],
         )
         gui = Invoke(
-            lambda txt: True if txt == 'True' else False,
+            __bool_from_string,
             LaunchConfiguration('gui'),
         )
 
@@ -312,6 +325,208 @@ def gz_spawn_entity(
         name,
         model_path,
         timeout_ms,
+    ).and_then(
+        __log_then_forward_cmd
+    ).and_then_with_key(
+        'cmd',
+        ExecuteProcess,
+    )
+
+
+class WorldReset(Flag):
+    """Todo."""
+
+    Time = auto()
+    Model = auto()
+    All = Time | Model
+
+
+def __make_world_control_cmd(
+        world: Text,
+        start: Optional[bool],
+        step: Optional[int],
+        reset: Optional[WorldReset],
+        seed: Optional[int],
+        timeout_ms: int,
+) -> Text:
+    if all(arg is None for arg in (start, step, seed, reset)):
+        raise AttributeError(
+            'No arguments provided to gz_control (Require at least one of'
+            ' "start", "step", "seed" and/or "reset" argument'
+        )
+
+    msg = 'On world "{}":'.format(world)
+
+    cmd = [
+        'gz',
+        'service',
+        '-s', '/world/{}/control'.format(world),
+        '--reqtype', 'gz.msgs.WorldControl',
+        '--reptype', 'gz.msgs.Boolean',
+        '--timeout', '{}'.format(timeout_ms),
+        '--req',
+    ]
+
+    requests = []
+
+    if start is not None:
+        msg += '\n - {}'.format('Starting sim' if start else 'Pausing sim')
+        requests.append('pause: {}'.format('false' if start else 'true'))
+
+    if step is not None:
+        if step == 1:
+            msg += '\n - Stepping'
+            requests.append('step: true')
+        elif step > 1:
+            msg += '\n - Stepping {} times'.format(step)
+            requests.append('multi_step: {}'.format(step))
+        else:
+            raise AttributeError(
+                '"step" argument MUST be >= 1 (step = {})'.format(step)
+            )
+
+    if reset is not None:
+        msg += '\n - Reseting ({})'.format(reset)
+        requests.append(
+            'reset: {{{}: true}}'.format(
+                'model_only' if reset == WorldReset.Model
+                else 'time_only' if reset == WorldReset.Time
+                else 'all'
+            )
+        )
+
+    if seed is not None:
+        if seed < 0:
+            raise AttributeError(
+                '"seed" argument MUST be >= 0 (seed = {})'.format(seed)
+            )
+
+        msg += '\n - Setting seed to {}'.format(seed)
+        requests.append(
+            'seed: {}'.format(seed)
+        )
+
+    logger.info(msg)
+
+    cmd.append(', '.join(requests))
+    return cmd
+
+
+def gz_control(
+        *,
+        world: Optional[SubstitutionOr[Text]] = None,
+        start: Optional[SubstitutionOr[bool]] = None,
+        step: Optional[SubstitutionOr[int]] = None,
+        reset: Optional[SubstitutionOr[WorldReset]] = None,
+        seed: Optional[SubstitutionOr[int]] = None,
+        timeout_ms: Optional[SubstitutionOr[int]] = None,
+) -> Generator[Action]:
+    """Spawn a model, with a given name, into an already running GZ server.
+
+    Parameters
+    ----------
+    world: Optional[SubstitutionOr[Path]]
+      If not None, the GZ world we wish to spawn our model into.
+      When None, declare a LaunchArgument for it (default to 'empty').
+    start: Optional[SubstitutionOr[bool]]
+      If not None, set to true if you wish to start the sim, false to pause.
+      When None, declare a LaunchArgument for it (default to '' - unset).
+    step: Optional[SubstitutionOr[uint]]
+      If not None, set to the number of steps you wish to do.
+      When None, declare a LaunchArgument for it (default to '' - unset).
+    reset: Optional[SubstitutionOr[WorldReset]]
+      If not None, set to one of the WorldReset flag (Time, Model or All).
+      When None, declare a LaunchArgument for it (default to '' - unset).
+    seed: Optional[SubstitutionOr[uint]]
+      If not None, set the random seed used by the world.
+      When None, declare a LaunchArgument for it (default to '' - unset).
+    timeout: Optional[SubstitutionOr[int]]
+      If not None, the timeout in ms associated to the gz service request.
+      When None, declare a LaunchArgument for it (default to 1000).
+
+    Returns
+    -------
+    Generator[Action]
+      All ROS launch Actions used to perform the needed task.
+    """
+    if world is None:
+        yield DeclareLaunchArgument(
+            'world',
+            description=(
+                'Name of the world we wish to spawn the entity into'
+            ),
+            default_value='empty',
+        )
+        world = LaunchConfiguration('world')
+
+    if start is None:
+        yield DeclareLaunchArgument(
+            'start',
+            description=(
+                'Set to "True" if you wish to start the sim. "False" to pause'
+                ' it.'
+            ),
+            default_value='',
+            choices=['False', 'True', ''],
+        )
+        start = Invoke(
+            __bool_from_string,
+            LaunchConfiguration('start'),
+        )
+
+    if step is None:
+        yield DeclareLaunchArgument(
+            'step',
+            description='Set to a number of steps to jump to (MUST BE >= 1)',
+            default_value='',
+        )
+        step = Invoke(
+            lambda t: None if t == '' else int(t),
+            LaunchConfiguration('step'),
+        )
+
+    if reset is None:
+        yield DeclareLaunchArgument(
+            'reset',
+            description='Reset either the Time or the Model (or both).',
+            choices=['', 'Time', 'Model', 'All'],
+            default_value='',
+        )
+        reset = Invoke(
+            lambda t: None if t == '' else WorldReset[t],
+            LaunchConfiguration('reset'),
+        )
+
+    if seed is None:
+        yield DeclareLaunchArgument(
+            'seed',
+            description='The random seed used by the model. (MUST BE >= 0)',
+            default_value='',
+        )
+        seed = Invoke(
+            lambda t: None if t == '' else int(t),
+            LaunchConfiguration('seed'),
+        )
+
+    if timeout_ms is None:
+        yield DeclareLaunchArgument(
+            'timeout_ms',
+            description='Timeout associated to the gz request (in ms)',
+            default_value='1000',
+        )
+        timeout_ms = Invoke(
+            int,
+            LaunchConfiguration('timeout_ms'),
+        )
+
+    yield Invoke(
+        __make_world_control_cmd,
+        world=world,
+        start=start,
+        step=step,
+        reset=reset,
+        seed=seed,
+        timeout_ms=timeout_ms,
     ).and_then(
         __log_then_forward_cmd
     ).and_then_with_key(