+++ CHANGELOG.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `lfc_parameters.yaml`: Match with the new parameters of  LFC v2.0.0
 
+### Added
+
+- `gz_control()`/`gz_control.launch.py`: New launch file use to control the sim
+  (start/pause/...) by cmd line
+
 ## [1.0.2] - 2025-04-24
 
 ### Fixed